Move modal logic to separate component
